build: load presets from unbundled `nitropack/presets`
