+++ src/presets/index.d.ts
@@ -1,3 +1,2 @@
-export { default as presets } from "./_all.gen"
-
 export type { PresetOptions, PresetName, PresetNameInput } from "./_types.gen";
+export { resolvePreset } from './_resolve'
