+++ src/presets/index.mjs
@@ -1,0 +1,1 @@
+export { resolvePreset } from './_resolve'
